refactor: Search domain 디렉토리 생성 및 관련 파일 이동
