+++ src/domains/Search/Page/Search.style.ts
@@ -10,9 +10,3 @@ export const SearchInput = styled.input`
     outline: none;
   }
 `;
-export const SearchListContainer = styled.ul`
-  height: calc(100vh - (var(--header-height)));
-  border-top: 1px solid var(--color-gray-300);
-
-  overflow-y: auto;
-`;
